+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Adminer dev
-- Tables overview: speed up
+- Tables overview: speed up, show totals in SQLite
 
 ## Adminer 5.5.1 (released 2026-07-21)
 - Ignore invalid X-Forwarded-Prefix (GHSA-fr74-9mf9-gf44)